Allow subsecond connect_timeout
Fixes #23

+++ lib/amqp/client/connection.rb
@@ -18,7 +18,7 @@ module AMQP
       # @option options [Boolean] connection_name (PROGRAM_NAME) Set a name for the connection to be able to identify
       #   the client from the broker
       # @option options [Boolean] verify_peer (true) Verify broker's TLS certificate, set to false for self-signed certs
-      # @option options [Integer] connect_timeout (30) TCP connection timeout
+      # @option options [Float] connect_timeout (30) TCP connection timeout
       # @option options [Integer] heartbeat (0) Heartbeat timeout, defaults to 0 and relies on TCP keepalive instead
       # @option options [Integer] frame_max (131_072) Maximum frame size,
       #    the smallest of the client's and the broker's values will be used
@@ -414,7 +414,7 @@ module AMQP
       # @return [Socket]
       # @return [OpenSSL::SSL::SSLSocket]
       def open_socket(host, port, tls, options)
-        connect_timeout = options.fetch(:connect_timeout, 30).to_i
+        connect_timeout = options.fetch(:connect_timeout, 30).to_f
         socket = Socket.tcp host, port, connect_timeout: connect_timeout
         keepalive = options.fetch(:keepalive, "").split(":", 3).map!(&:to_i)
         enable_tcp_keepalive(socket, *keepalive)
